graph/requirements.txt
fastapi==0.97.0
pydantic==1.10.9
uvicorn==0.22.0
strawberry-graphql[fastapi]==0.184.0
aiosqlite==0.19.0
sqlalchemy==2.0.16
strawberry-sqlalchemy-mapper==0.1.1
psycopg2-binary==2.9.6
Pillow==9.5.0
numpy==1.24.3
CairoSVG==2.7.0